Add parent class _wood
