+++ smolyay/benchmark/s.py
@@ -9,6 +9,7 @@ from .kow import _kow
 from .osborne import _osborne
 from .powell import _powell
 from .rosenbr import _rosenbrock
+from .wood import _wood
 
 
 class s201(BenchmarkFunction):
@@ -411,42 +412,10 @@ class s257(BenchmarkFunction):
         return v[...,1]
 
 
-class s258(BenchmarkFunction):
+class s258(_wood):
     @property
     def domain(self):
         return [[-9.0, 9.9], [-9.0, 9.9], [-9.0, 9.9], [-8.9999999999, 9.90000000009]]
-
-    def _function(self,x):
-        v = numpy.zeros((x[...,0].size,4))
-        v[...,0] = x[...,0] * x[...,0]
-        v[...,1] = x[...,1] - v[...,0]
-        v[...,0] = v[...,1] * v[...,1]
-        v[...,1] = 100. * v[...,0]
-        v[...,0] = 1. - x[...,0]
-        v[...,2] = v[...,0] * v[...,0]
-        v[...,1] += v[...,2]
-        v[...,2] = x[...,2] * x[...,2]
-        v[...,0] = x[...,3] - v[...,2]
-        v[...,2] = v[...,0] * v[...,0]
-        v[...,0] = 90. * v[...,2]
-        v[...,1] += v[...,0]
-        v[...,0] = 1. - x[...,2]
-        v[...,2] = v[...,0] * v[...,0]
-        v[...,1] += v[...,2]
-        v[...,2] = -1. + x[...,1]
-        v[...,0] = v[...,2] * v[...,2]
-        v[...,2] = 10.1 * v[...,0]
-        v[...,1] += v[...,2]
-        v[...,2] = -1. + x[...,3]
-        v[...,0] = v[...,2] * v[...,2]
-        v[...,2] = 10.1 * v[...,0]
-        v[...,1] += v[...,2]
-        v[...,2] = -1. + x[...,1]
-        v[...,0] = 19.8 * v[...,2]
-        v[...,2] = -1. + x[...,3]
-        v[...,3] = v[...,0] * v[...,2]
-        v[...,1] += v[...,3]
-        return v[...,1]
 
 class s259(BenchmarkFunction):
     @property
