+++ smolyay/benchmark/wood.py
@@ -1,0 +1,36 @@
+import numpy
+
+from .benchmark import BenchmarkFunction
+
+class _wood(BenchmarkFunction):
+    def _function(self,x):
+        v = numpy.zeros((x[...,0].size,4))
+        v[...,0] = x[...,0] * x[...,0]
+        v[...,1] = x[...,1] - v[...,0]
+        v[...,0] = v[...,1] * v[...,1]
+        v[...,1] = 100. * v[...,0]
+        v[...,0] = 1. - x[...,0]
+        v[...,2] = v[...,0] * v[...,0]
+        v[...,1] += v[...,2]
+        v[...,2] = x[...,2] * x[...,2]
+        v[...,0] = x[...,3] - v[...,2]
+        v[...,2] = v[...,0] * v[...,0]
+        v[...,0] = 90. * v[...,2]
+        v[...,1] += v[...,0]
+        v[...,0] = 1. - x[...,2]
+        v[...,2] = v[...,0] * v[...,0]
+        v[...,1] += v[...,2]
+        v[...,2] = -1. + x[...,1]
+        v[...,0] = v[...,2] * v[...,2]
+        v[...,2] = 10.1 * v[...,0]
+        v[...,1] += v[...,2]
+        v[...,2] = -1. + x[...,3]
+        v[...,0] = v[...,2] * v[...,2]
+        v[...,2] = 10.1 * v[...,0]
+        v[...,1] += v[...,2]
+        v[...,2] = -1. + x[...,1]
+        v[...,0] = 19.8 * v[...,2]
+        v[...,2] = -1. + x[...,3]
+        v[...,3] = v[...,0] * v[...,2]
+        v[...,1] += v[...,3]
+        return v[...,1]
